refactor(avatar): 优化目录结构
