+++ packages/devui-vue/devui/avatar/src/components/avatar-body-icon.tsx
