+++ packages/devui-vue/devui/avatar/src/components/avatar-nobody-icon.tsx
