+++ packages/devui-vue/devui/avatar/src/components/icon-nobody.tsx
@@ -1,4 +1,4 @@
-import type { IconPropsType } from "./avatar-types";
+import type { IconPropsType } from "../avatar-types";
 
 export const IconNobody = (props: IconPropsType): JSX.Element => {
   const { width, height } = props;
